improv: threads and posts perf

+++ src/models/models.go
@@ -19,29 +19,29 @@ type UserUpdate struct {
 }
 
 type Forum struct {
-	Title string `json:"title"`
-	User string `json:"user"`
-	Slug string `json:"slug"`
-	Posts int `json:"posts"`
-	Threads int `json:"threads"`
+	Title   string `json:"title"`
+	User    string `json:"user"`
+	Slug    string `json:"slug"`
+	Posts   int    `json:"posts"`
+	Threads int    `json:"threads"`
 }
 
 type ServiceStatus struct {
-	UserCount int `json:"user"`
-	ForumCount int `json:"forum"`
+	UserCount   int `json:"user"`
+	ForumCount  int `json:"forum"`
 	ThreadCount int `json:"thread"`
-	PostCount int `json:"post"`
+	PostCount   int `json:"post"`
 }
 
 type Thread struct {
-	Id      int    `json:"id"`
-	Forum   string `json:"forum"`
-	Title   string `json:"title"`
-	Author  string `json:"author"`
-	Message string `json:"message"`
+	Id      int       `json:"id"`
+	Forum   string    `json:"forum"`
+	Title   string    `json:"title"`
+	Author  string    `json:"author"`
+	Message string    `json:"message"`
 	Created time.Time `json:"created"`
-	Votes   int64    `json:"votes"`
-	Slug    string `json:"slug"`
+	Votes   int       `json:"votes"`
+	Slug    string    `json:"slug"`
 }
 
 type ThreadUpdate struct {
@@ -50,14 +50,14 @@ type ThreadUpdate struct {
 }
 
 type Post struct {
-	Id int `json:"id"`
-	Parent int `json:"parent"`
-	Author string `json:"author"`
-	Message string `json:"message"`
-	IsEdited bool `json:"isEdited"`
-	Forum string `json:"forum"`
-	Thread int `json:"thread"`
-	Created time.Time `json:"created"`
+	Id       int       `json:"id"`
+	Parent   int       `json:"parent"`
+	Author   string    `json:"author"`
+	Message  string    `json:"message"`
+	IsEdited bool      `json:"isEdited"`
+	Forum    string    `json:"forum"`
+	Thread   int       `json:"thread"`
+	Created  time.Time `json:"created"`
 }
 
 type PostUpdate struct {
@@ -65,15 +65,15 @@ type PostUpdate struct {
 }
 
 type PostDetails struct {
-	Post Post `json:"post"`
+	Post   Post   `json:"post"`
 	Thread Thread `json:"thread"`
-	Forum Forum `json:"forum"`
-	User User `json:"user"`
+	Forum  Forum  `json:"forum"`
+	User   User   `json:"user"`
 }
 
 type ThreadVote struct {
-	Id int `json:"id"`
-	Thread int `json:"thread"`
+	Id       int    `json:"id"`
+	Thread   int    `json:"thread"`
 	Nickname string `json:"nickname"`
-	Voice int64 `json:"voice"`
+	Voice    int    `json:"voice"`
 }
